Follow the pagination in the repositories listing
Registry#search only fetched the first page, even though the API specification
states that the catalog route can be paginated. Let’s fetch all the pages
instead.

The function prototype was kept simplistic, unlike Registry#tag which provides
finer controls but is much harder to use. The rationale is that most users
expect the full list and do not have to be aware of the pagination details of
the API. If the eager loading causes issues, we could return a lazy Enumerator.